[dart2js] Clear source location mappings after they've been processed.
The data within this maps is apparently hanging around in memory longer than it needs to. We only read from each map in a short window and once it's been read we don't use that data again. By clearing the maps ater we use them we end up saving ~700MB (out of ~7.1GB) for phase 3b on a large program.

References to all these maps are being retained by this higher level map:
https://github.com/dart-lang/sdk/blob/main/pkg/compiler/lib/src/io/code_output.dart#L143

This change clears the internal maps while that larger map is being iterated over.

Change-Id: I132dbf469391f257685b05fa69c302f1f1c44e31
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/325940
Reviewed-by: Mayank Patke <fishythefish@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>