[vm/compiler] In CP ignore constant Redefinitions
Only replace Redefinition with a constant if the
original definition is a constant.

It used to be the case that Redefinition could
be constant only if the origin definition was
however 5c320a108 changed that: we would augment
graph with additional redefinitions after branches
on equlity comparisons. This meant that constant
value would propagate to other already inserted
redefinitions and than these redefinitions started
to be replaced by their constant value. Something
that we wanted to prevent to begin with.

Also fix the comment for InsertRedefinitionsAfterEqualityComparisons
as pointed by @aam

Fixes https://github.com/dart-lang/sdk/issues/53514

R=alexmarkov@google.com
TEST=issue found by CI and so covered by existing tests

Fixed: 53514
Change-Id: I5e547d34de3aabc340329da1a43341689bfc58e1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/326120
Reviewed-by: Alexander Markov <alexmarkov@google.com>
Auto-Submit: Slava Egorov <vegorov@google.com>
Commit-Queue: Slava Egorov <vegorov@google.com>