[beta] Fix type promotion of late final fields on DDC and dart2wasm.
When the CFE compiles code for DDC or dart2wasm, it lowers late final
fields to synthetic getters. In order to ensure that accesses to these
fields can still be type promoted, a flag is added to the kernel
`Procedure` class, to indicate that a getter was produced by lowering
a late final field. When deciding whether a property access is
promotable, the CFE checks this flag; if it's set, it treats the
getter like a field.

Includes a language test validating the fix.

TEST=tests/language/inference_update_2/late_field_test.dart

Bug: https://github.com/dart-lang/sdk/issues/53462
Change-Id: I6d15fe13201d4c8fb7a3e9f81b9a937092cc86e3
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/324767
Cherry-pick-request: https://github.com/dart-lang/sdk/issues/53490
Fixes: https://github.com/dart-lang/sdk/issues/53490
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/325142
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>