[stable] [dart2js] Erase static interop type in static invocation
Static invocations of external factories are casted so that the
result, which is a @staticInterop type, can be treated as the erased
type instead. This CL fixes the issue where the type that it was
casted to was never replaced with the erased type.

Change-Id: Ic6eb529349ea2b5c42f91c2740d501d4f81bc38e
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/323505
Cherry-pick-request: https://github.com/dart-lang/sdk/issues/53579
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/327120
Reviewed-by: Sigmund Cherem <sigmund@google.com>