[analyzer] Fix errors for static object record fields
The analyzer reports an error when the name of a field in a record is
already defined on the `Object` class. This is slightly off the
intended behavior - the specification actually has a fixed list of
forbidden names. In particular, the mismatch was observed when using
names of static members on the `Object` class as record fields.

Closes https://github.com/dart-lang/sdk/issues/53257

Change-Id: Ib859b2acdc8e5b1917f3859926cbd87579067048
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/327580
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>