Analyzer: only call FlowAnalysis.propertyGet for instance getters.
When a syntax that looks like a method invocation syntax actually
winds up resolving to a getter rather than a method, the analyzer
rewrites it into a function expression invocation. The getter in
question could be an instance getter, a static getter, a top level
getter, or a read of a local variable.

Previously, the analyzer was calling `FlowAnalysis.propertyGet` in all
of these circumstances. But this method should only be used for
invocations of instance getters.

The situation is benign right now, but in an upcoming CL, the
functionality of `FlowAnalysis.propertyGet` will be expanded to track
non-promotion reasons for property gets. To avoid confusion when that
happens, we need to ensure that `FlowAnalysis.propertyGet` is only
called for things that are genuine property gets.

Change-Id: I400826d5fd5fa74ced5c005476c8617252a73cae
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/327900
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>