Handle extension types in front end field promotion logic.
The front end has logic in
`SourceLibraryBuilder.computeFieldPromotability` to add entries to
`FieldNonPromotabilityInfo.individualPropertyReasons` for getters that
appear in extensions, to record that they are not promotable because
they are getters. The same logic needs to be replocated for getters
that appear in extension types.

Change-Id: Ie6796ed507ae6b3c1190621f1b03e62e1e4067f0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/328060
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>