Add information about field promotion conflicts to element model.
There are two ways an access to a private property name might not be
promotable: the property might be non-promotable for a reason inherent
to itself (e.g. it's declared as a concrete getter rather than a
field, or it's a non-final field), or the property might have the same
name as an inherently non-promotable property elsewhere in the same
library (in which case the inherently non-promotable property is said
to be "conflicting").

When a compile-time error occurs because a property is non-promotable
due conflicting properties elsewhere in the library, the analyzer
needs to be able to find the conflicting properties in order to
generate context messages. This CL adds the necessary data to the
summary format so that the analyzer will be able to map each private
name to the set of conflicting declarations.

The functionality of generating the appropriate context messages will
be added in a follow-up CL.

Change-Id: I29df6f562b3495300b01be1edbc80b9a0f684df5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/327904
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>