Analyzer: "why not promoted" messages for failed field promotion due to conflict.
During summary generation for a library, when
`_FieldPromotability.perform` computes which fields are promotable,
the data computed by the shared method `computeNonPromotabilityInfo`
is serialized into the element model, in
`LibraryElementImpl.fieldNameNonPromotabilityInfo`. This data reports,
for each private getter name, precisely which declarations it
conflicts with. These declarations are explicit fields, explicit
getters, and classes that contain implicit `noSuchMethod`-forwarding
getters.

Later, if a compile-time error results from a failure in field
promotion due to a conflict between the field in question and another
field or getter of the same name elsewhere in the library,
`_WhyNotPromotedVisitor.visitPropertyNotPromotedDueToConflict` is
called. It iterates through the information stored in
`LibraryElementImpl.fieldNameNonPromotabilityInfo` and generates
context messages for each conflicting declaration.

Change-Id: I8f63425384c917db9bbc6e6dd5990a3d10a3a7ff
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/328061
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Nate Bosch <nbosch@google.com>