[analyzer] Consider VarArgs in `@Native` check
Methods and functions annotated with `@FfiNative` and `@Native` are
validated to ensure their Dart signature matches the one from the
annotation.
For those functions, the lengths of the parameters is compared two
times: In `_validateCompatibleFunctionTypes` and in `_checkFfiNative`,
with the latter being responsible for emitting a more-specific error
message. However, it fails to consider that a single `VarArg` reference
in the native function type would get expanded to multiple formals in
Dart. This fixes the problem by properly expanding variadic types in
`_checkFfiNative` as well.

Closes https://github.com/dart-lang/sdk/issues/53680

Bug: 53680
Change-Id: Ifb93eba8ac9f39697dc415dc337110109449e366
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/329360
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Daco Harkes <dacoharkes@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>