[beta][vm] Enable waitFor for 3.2 release
We are going to enable waitFor in 3.2 release
and disable it in 3.3.

This makes migration smoother for users that need
to use Isolate.resolvePackageUriSync which is
only added in 3.2.

The actual removal timeline remains the same:
we will remove waitFor in 3.4.

TEST=ci

Bug: https://github.com/dart-lang/sdk/issues/52121
Change-Id: If38376c66e1d6712c3d030a1b4c88f1ab121c204
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/329420
Cherry-pick-request: https://github.com/dart-lang/sdk/issues/53691
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/329560
Reviewed-by: Martin Kustermann <kustermann@google.com>