add a `hasImmutable` property to elements
For whatever reason, we (probably I) neglected to add this property and we're fiddling with metadata directly in the linter to find out if an `@immutable` annotation exists on declarations.

This just gets us consistent and will allow me to tidy up a lint or two.

Change-Id: I398d7527227a0407387bd45c474d59b5185c413d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/330048
Commit-Queue: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Kallen Tu <kallentu@google.com>